+++ NobatPlusDATA/Domain/StylistService.cs
@@ -13,7 +13,7 @@ namespace NobatPlusDATA.Domain
 
         public long ServiceManagementID { get; set; }
         public ServiceManagement ServiceManagement { get; set; }
-       public long ServicePrice { get; set; }
+       public decimal ServicePrice { get; set; }
         public TimeSpan ServiceDuration { get; set; }
         public int DepositPercent { get; set; }
     }
